Added example for PrincipalOrgID global condition key
When using a domain with many accounts, it may be easier to authorize the entire organization instead of individual account principals.  Using the global condition key is a way to do this.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the terms of your choice.